DRAW-287 refactor: 레디스 마피아 게임 정보 저장 클래스 분리
